# desktop/src/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Recetario API",
        "version": "0.1.0"
    },
    "paths": {
...
        "/ingestion/jobs": {
            "get": {
                "tags": [
                    "ingestion"
                ],
                "summary": "List Ingestion Jobs",
                "operationId": "list_ingestion_jobs_ingestion_jobs_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/IngestionJobOut"
                                    },
                                    "type": "array",
                                    "title": "Response List Ingestion Jobs Ingestion Jobs Get"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "ingestion"
                ],
                "summary": "Create Ingestion Job",
                "operationId": "create_ingestion_job_ingestion_jobs_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IngestionJobCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IngestionJobOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/ingestion/jobs/{job_id}": {
            "get": {
                "tags": [
                    "ingestion"
                ],
                "summary": "Get Ingestion Job",
                "operationId": "get_ingestion_job_ingestion_jobs__job_id__get",
                "parameters": [
                    {
                        "name": "job_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Job Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IngestionJobOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "IngestionInputType": {
                "type": "string",
                "enum": [
                    "web",
                    "video"
                ],
                "title": "IngestionInputType",
                "description": "How the source should be fetched. Only WEB is wired in this slice."
            },
            "IngestionJobCreate": {
                "properties": {
                    "url": {
                        "type": "string",
                        "minLength": 1,
                        "title": "Url",
                        "description": "Recipe page URL to import."
                    }
                },
                "type": "object",
                "required": [
                    "url"
                ],
                "title": "IngestionJobCreate"
            },
            "IngestionJobOut": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "input_url": {
                        "type": "string",
                        "title": "Input Url"
                    },
                    "input_type": {
                        "$ref": "#/components/schemas/IngestionInputType"
                    },
                    "status": {
                        "$ref": "#/components/schemas/JobStatus"
                    },
                    "progress": {
                        "type": "integer",
                        "title": "Progress"
                    },
                    "result_recipe_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Result Recipe Id"
                    },
                    "error": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Error"
                    },
                    "created_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Created At"
                    },
                    "updated_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Updated At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "input_url",
                    "input_type",
                    "status",
                    "progress",
                    "result_recipe_id",
                    "error",
                    "created_at",
                    "updated_at"
                ],
                "title": "IngestionJobOut"
            },
...
            "JobStatus": {
                "type": "string",
                "enum": [
                    "queued",
                    "running",
                    "succeeded",
                    "failed"
                ],
                "title": "JobStatus"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}